Refactor methods for UI
